FIND BUG! uint16_t adcBuffer[ADC_BUFFER_SIZE]; 未处理DMA对齐问题：
你使用的是uint32_t adcBuffer[]来存储ADC采集的数据，而DMA配置的是16位对齐。建议将adcBuffer定义为uint16_t类型

+++ Core/Src/main.c
@@ -49,7 +49,7 @@
 
 /* USER CODE BEGIN PV */
 #define ADC_BUFFER_SIZE 77
-uint32_t adcBuffer[ADC_BUFFER_SIZE];
+uint16_t adcBuffer[ADC_BUFFER_SIZE];
 
 //uint32_t ADC_Value[ADC_BUFFER_SIZE]; // 声明数组来存储ADC采样结果
 //uint16_t filtered_adc_values[7]; // 过滤后的ADC
